\name{S}
\alias{S}
\docType{data}
\title{
  ER dataset
}
\description{
  Empirical covariance matrix derived from the ER dataset.  The original
  dimension 7027 was reduced to 692 by thresholding. 
}
\source{
  http://www.math.nus.edu.sg/~mattohkc/Covsel-0.zip
}
\references{
  Lu Li, Kim-Chuan Toh: An inexact interior point method for L1 regularized sp\
arse covariance selection.  Math. Prog. Comp. (2010) 2:291-315
}
\keyword{datasets}
